validation.goに既存のvalidationを移行
